Added external logging support. Refactory namespace to 'com.bigludo'.
