--- src/main/java/ILog.java
@@ -1,7 +1,0 @@
-package main.java;
-
-public interface ILog {
-    void info (String message);
-
-    void debug (String message);
-}
