+++ src/main/java/com/bigludo/IClient.java
@@ -1,4 +1,4 @@
-package main.java;
+package com.bigludo;
 
 public interface IClient {
