+++ src/main/java/com/bigludo/IServer.java
@@ -1,4 +1,4 @@
-package main.java;
+package com.bigludo;
 
 public interface IServer {
