+++ src/main/java/com/bigludo/ServerHandler.java
@@ -1,15 +1,14 @@
-package main.java;
+package com.bigludo;
 
 import java.util.ArrayList;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class ServerHandler implements IServer{
     private ArrayList <IClient> clients = new ArrayList<>();
-    private ILog log;
 
-    public ServerHandler (ILog lg) {
-        log = lg;
-    }
-
+    private static final Logger log = LogManager.getLogger(ServerHandler.class);
+    
     public void register (IClient client) {
         clients.add(client);
         log.info("Register client " + client.getId());
